refreshed script for detecting orphaned credentials
