+++ src/orphaned-credentials/users-credentials.graphql
@@ -1,0 +1,14 @@
+query usersWithCredentials {
+  users {
+    id
+    nameID
+    agent {
+      id
+      credentials {
+        id
+        type
+        resourceID
+      }
+    }
+  }
+}
